requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
pandas
scikit-learn
numpy
python-dotenv
